requirements.txt
absl-py==0.9.0
astor==0.8.1
-e git+https://github.com/openai/baselines.git@ea25b9e8b234e6ee1bca43083f8f3cf974143998#egg=baselines
click==7.1.2
cloudpickle==1.2.0
future==0.18.2
gast==0.3.3
grpcio==1.29.0
gym==0.15.7
h5py==2.10.0
importlib-metadata==1.6.1
joblib==0.15.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.18.5
opencv-python==4.2.0.34
protobuf==3.12.2
pyglet==1.5.0
PyYAML==5.3.1
scipy==1.4.1
six==1.15.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
tqdm==4.46.1
Werkzeug==1.0.1
zipp==3.1.0
